Cache generated structs
